move all to tests folder
